Implement read or generate index from CAR accepting v1 or v2
A user has to undergo a lot of boilerplate code to generate or get an
index where there is a CAR file, the version for which is unknown.
Provide a function that accepts either v1 or v2 payload and reads the
index if present, otherwise generates it for the user. The generated
index lives entirely in memory and will not depend on the given
`io.ReadSeeker` to lookup indices.

Implement tests that check index is returned for v1 or v2 files. Add a
sample file with an imaginary version of 42 for testing purposes to
assert that the indexing function will error if the given file is not v1
or v2.

See: https://github.com/ipld/go-car/issues/143


This commit was moved from ipld/go-car@fc494d46623af156f057ef363c1ee2a7ffe0cb13