Avoid creating files on resumption in`ReadWrite` blockstore
When resumption is enabled, we do not want to create a file if it does
not exist. Similarly, when resumption is disabled, fail if the given
file at path already exists to avoid unintended file overrides and
creations.


This commit was moved from ipld/go-car@8c08fbfef55c8f1463bbd4eb3004011bc076a23a