Avoid writing to files in testdata
Avoid writing to files in `testdata` through examples. This is because
when tests that use those file are run in parallel will fail if files
are modified. Instead write to a temporary file in examples, and
whenever opening testdata files in `RW` mode, make a temporary copy.

Thanks to @mvdan for pointing this out.

Fixes #175


This commit was moved from ipld/go-car@c3a595560230886d34e35e63ce258eb79c815ccc