Propagate async `blockstore.AllKeysChan` errors via context
Implement a mechanism that allows the user to hook an error handling
function to the context passed to `AllKeysChan`. The function is then
notified when an error occurs during asynchronous traversal of data
payload.

Add tests that assert the success and failure cases when error handler
is set.

Fixes #177


This commit was moved from ipld/go-car@de2711c09ab3267aba84f7b8bd8ecaaeb4885f21