update package godocs and root level README for v2
While at it, make v2's license a symlink, so the two stay in sync.
The reason we want the v2 module to have its own license is so that its
module zip includes the file as well.

Besides mentioning v0 and v2, the root README now also links to pkgsite
and lists Masih and myself as default maintainers, given that we're the
ones that did major work on this library last.

Fixes #124.
Fixes #179.


This commit was moved from ipld/go-car@61cfb4dbcfc4bf48c54425385a1aae279b2b6ca1