avoid another alloc per read byte
Like the last commit, avoid an extra allocation per read byte.
In this case, we created a one-byte buffer for each readByte call.
Unfortunately, since io.Reader is an interface,
the compiler can't know if it holds onto the memory,
so the buffer escapes and cannot be placed in the stack.

To sidestep this issue, reuse a preallocated buffer.
We know this is fine, because we only do sequential reads.

    name           old time/op    new time/op    delta
    ReadBlocks-16    1.15ms ± 5%    1.09ms ± 4%   -5.13%  (p=0.000 n=11+11)

    name           old speed      new speed      delta
    ReadBlocks-16   453MB/s ± 5%   478MB/s ± 4%   +5.41%  (p=0.000 n=11+11)

    name           old alloc/op   new alloc/op   delta
    ReadBlocks-16    1.29MB ± 0%    1.30MB ± 0%   +0.48%  (p=0.000 n=12+12)

    name           old allocs/op  new allocs/op  delta
    ReadBlocks-16     11.5k ± 0%      9.5k ± 0%  -17.35%  (p=0.000 n=12+12)


This commit was moved from ipld/go-car@ccffb5c06ed2c79d67dba037a4754b4212480790