blockstore: OpenReadWrite should not modify if it refuses to resume
If OpenReadWrite sees a good reason to refuse to resume,
such as the provided roots not matching what's on disk,
it should not truncate or un-finalize the file on disk.

Add a test that exercises that edge case,
and also verifies that the file isn't modified.

To fix the problem, carefully move truncation and un-finalization
past the point where we've checked the existing header.

Note that we may still leave a broken file if we encounter an error
later on, such as when writing to the file or building the index.
If we want those to not break the input file we'll need bigger changes,
such as writing to a temporary file and atomically replacing the final
file at the very end.

For now, at least we can avoid a common breakage case.

Before the fix, the test would fail:

    Error Trace: readwrite_test.go:621
    Error:       Not equal:
                   expected: 521708
                   actual  : 479958
    Test:        TestReadWriteResumptionMismatchingRootsIsError

Fixes #247.


This commit was moved from ipld/go-car@4e0a1fa04c8f7e8fc69176007bc07c25ef7ce9c9