Remove support for `ForEach` enumeration from car-index-sorted
This index type does not store enough information to satisfy `ForEach`.
It only contains the digest of mulithashes and not their code. Instead
of some partial functionality simply return an error when
`ForEach` is called on this function type. Because, there is no valid
use for this index type and the user should ber regenerating the index
to the newer `car-multihash-index-sorted` anyway.

Update tests to include samples of both types and assert IO operations
and index generation for both formats.


This commit was moved from ipld/go-car@5c7e07edac3835e83ee3452507a9ba3bc94f8d2e