feat: Has() and Get() will respect StoreIdentityCIDs option
When StoreIdentityCIDs is set, it will defer to the index to check whether the
blocks are in the CAR. When the CAR is a v1 and StoreIdentityCIDs is set, the
index will contain the identity CIDs. When it's a v2 with an existing index,
however that index was created will determine whether the identity CIDs are
that are in the CAR are found.

When StoreIdentityCIDs is not set, Has() will always return true and Get() will
always return the block.


This commit was moved from ipld/go-car@02d658faa7dfbc01920fc42ad7210c3505d049f4