fix(gateway): JSON when Accept is a list
Block/CAR responses always had single explicit type, and we did not bother
with implementing/testing lists.

With the introduction of JSON people may start passing a list.
This is the most basic fix which will return on the first matching
type (in order). This does not implements weights (can be added in
future, if needed).

Closes #9520


This commit was moved from ipfs/kubo@b333740468fc6bd249878f8b82ef59a431879a6a