change the Pinner interface to have async pin listing
The rational is that if the pin list get big, a synchronous call to get the complete list can delay handling unnecessarily. For example, when listing indirect pins, you can start walking the DAGs immediately with the first recursive pin instead of waiting for the full list.

This matters even more on low power device, of if the pin list is stored remotely.

Replace:
- https://github.com/ipfs/go-ipfs-pinner/pull/24
- https://github.com/ipfs/go-ipfs-provider/pull/48