pinner: change the interface to have async pin listing
The rational is that if the pin list get big, a synchronous call to get the complete list can delay handling unnecessarily. For example, when listing indirect pins, you can start walking the DAGs immediately with the first recursive pin instead of waiting for the full list.

This matters even more on low power device, of if the pin list is stored remotely.

* coreiface: allow to return an error not linked to a specific Cid
* merkledag/test: add a DAG generator

Rationale is that generating a test DAG is quite difficult, and anything that helps writing better tests is helpful.