fix(routing/http): avoid escaping comma separator
The comma is in the "sub-delims" set of characters that don't need to be
encoded in most parts of a URL, including query parameters. Golang
standard library percent-escapes it for consistency, but we prefer
human-readable /routing/v1 URLs, and real comma is restored here to
ensure human and machine requests hit the same HTTP cache keys.