fix sending cancels when excluding peer
When sending cancels for want-block and want-have requests, a peer may be sxcluded because it was the sender of a block and that peer will cancel the want itself without having to receive a cancel message. The would-be sender of tha cancel message still needs to do everything, such as adjusting want counts, as if the cancel were being sent and then not send out the cancel message.