bitswap/httpnet: do not sleep on backoffs
Before: when a url is in cooldown, we sleep for the rest of the cooldown and
then proceed with the request.

Now: we return an error and try with the next url until no more to try.

The reason is we should not block a worker. We can avoid contacting servers during cooldowns, but we cannot schedule requests to be tried later or have workers waiting.

Scheduling is a footgun as it requires to leave some requests hanging for later and those queues will need to have a limit.

Blocking prevents usage of resources by others, and also prevents the message sender from returning as soon as possible with whatever results it obtained.