bitswap/httpnet: do not follow redirects
Avoid following redirects. We treat them as an incorrect provider record.  We
don't want to be coerced into opening new connections to new hosts and we
prefer provider records to be up to date rather than them relying in redirects
being followed. There are also problems with potential loops etc.

https://github.com/ipfs/boxo/issues/862