bitswap/httpnet: Improvements around peer Connect()
Harden the logic surrounding peer connections.

The NewMessageSender should only perform HTTP requests if we have gone through
Connect() and we are connected.

The router should failover to Bitswap when we are not connected even if we
somehow ended with some http addresses in the peerstore.

The Connect() logic is improved (try HEAD first, more efficient code...).