bitswap: add requestsInFlight metric
This metric mimics requestsInFlight in httpnet.

It should tell how many streams are being written or read at a given time,
just like in HTTP we measure how many requests are ongoing at a given time.

The main insight is knowing with how many "workers" (or how much concurrency) bitswap is using
for writing/reading wantlists to/from other hosts. It should roughly match the following, if both
things have the same performance:

bitswapRequestsInFlight = elementsInWantlist * httpRequestsInFlight

However, since httpRequestsInFlight is limited by the number of http workers, it may signal
that the number of workers is too low to compete with bitswap and that it should increase to
make better use of available concurrent stream capacities.