Improve performance of data onboarding (#888)
* Improve the perceived performance of data onboarding

Adding a large directory of data to IPFS can take a long time. A significant amount of this time is spent writing to and reading from the persisted provider queue, depending on the underlying datastore. The time spent doing this can be greatly reduced by writing queued items to the datastore in batches. In addition to batched writes, reading queued items can be read directly from the input buffer if there are no queued items in the datastore. This allows in-memory only operation when the number of queued items is below the batch size, and when the queue is not idle for too long.

* no ctx in struct
* ignore undef cids
* Rename queue.NewQueue to queue.New
* Keep CIDs in memory until batchSize reached. If nothing in datastore then read items to dequeue from memory buffer.