mfs: support MaxLinks and MaxHAMTFanout
Directory metadata like mode or mtime is serialized and added to the dag-pb
protobuf, so it is persisted with it and reloaded every time a unixfs folder
is loaded in into an MFS directory or cached as a child inside one.

With MaxLinks and MaxHAMTWidth we don't have this luxury. The idea is that,
if we have an MFS root, it is wrapping an unixfs dynamic directory, and at the same
time it is wrapping some protobuf (either basic or hamt).

Now, when we use it to operate, add subfolders, flush it, sync it etc, we must still
remember the max-links configuration and the only way we can do that is to set it
to what the parent has (while leaving Stat alone).

This should work when we are adding things to MFS and everything expects the
same options. When dealing with a long-lived MFS system and we import existing
DAGs into it, this is likely going to cause undesired effects (I'm unsure
about the current behavour but it may trigger HAMT conversions where before
things were left alone etc.). In any case, previous behavour is kept when the
options are zero'ed, which happens unless they are explicitally set (so we
should be safe wrt Kubo MFS DAG).

Implementation required public setters and getters in unixfs/io.Directory.

When adding files, for example it is a normal operation to create a new MFS
root.

Currently this requires creating an empty dag-pb node configured with
a Cid builder etc. and then using it to initialize a new root.

The new NewEmptyRoot() method exposes the unixfs directory options at the MFS
level. It simplifies making new roots and it lets us pass the unixfs options
that we wish, including setting MaxWidth or MaxHAMTFanout.