bitswap/httpnet: Disconnect peers after client errors
This adds a configuration option to disconnect from peers when they return too many errors of
type client (404s, Gone, Forbidden, Moved, Redirect. etc). in a row.

Such peers with low hit rates on the blocks we ask for must not perpetually
receive optimistic requests and are "disconnected" accordingly if they cross
the threshold (100 by default). They will be recontacted if a record pointing
to them appears.

The implementation logs the errors at the end of processing the responses for
a wantlist (that is, once per wantlist), the tracking only locks the map on
connect/disconnect events, allowing for low lock contention operations.