Do not return error from notify blocks when bitswap shutdown (#947)
On ipfs shutdown, MFS closes its root which adds an empty directory block. This results in an error because bitswap is shutdown, even though there is no real error to address.

To fix this, this PR changes NotifyNewBlocks to not return any error if bitswap is closed. Reporting this error is not necessary because there is not action that can or should be taken to handle it, and bitswap was closed intentionally so this is not even an unexpected situation.