bitswap/httpnet: limit metric cardinality
Currently, we label the StatusCounter metric with the hostname of the http
endpoint. If there was a high number of endpoints, this would cause high
cardinality for this metric.

This is a non-problem now as there are very few HTTP endpoints. This commit
addresses it anyways by only labelling the metric hostnames when the Allowlist
is enabled, therefore limiting labels to hostnames in the allowlist.