bitswap/httpnet: adjust error logging (#958)
Moved some error logging to warn level.

The errors are useful when figuring out why something is not loading from
somewhere (particularly in an scenario where an allowlist is configured), but
in an all-open setup they will spam the logs. Default log level is error I
think, so warn messages should not show and can be activated without getting
into the debug details.

Fixes #955.
Fixes #954.