Remove providing Exchange. Call Provide() from relevant places.
See https://github.com/ipfs/boxo/issues/963

Refactor the logic to perform Provides:
  * Remove providing Exchange()
  * Provide directly from Blockstore
  * Provide directly from pinner/merkledag on dag traversal
  * (Provides only happen when the component has been given a provider)

One problem I have is that of circular dependencies between the `provider`
module and the `blockstore/pinner/merkledag`. The main issue is that the
provide interfaces are not separate from the reprovider implementation. For
the moment I have copied the interfaces into the several modules to avoid
this.

MFS is untouched, for I believe we need to wrap the DagService as a "providing
DAG service", or build a DAGService backed by a providing blockstore. It seems
these can be done externally though.