bitswap/httpnet: fix peers silently stopping from doing http requests
Fixes #979 (hopefully). If connectivity status has been switched from
Connected without doing an httpnet DisconnectFrom(), it may be that we keep
pinging the peers, and think we are connected to them, but never restore the
connectivity status when skipping the connectivity checks (because we are
connected). This results in silent, successful "Connect()" requests that do
nothing.