bitswap/httpnet: start pinging before signaling Connected
Assuming that the stack might get a Connected notification,
but we have not yet started pinging.

Attempting to send a message at that moment fails, the messagequeue shutdown as a result, but no disconnect/connect happens from that point as it expects.

Also add some logging to see if we are hitting those message sender exists.