Filestore: add Provider option to provide filestore blocks. (#1042)
* Filestore: add Provider option to provide filestore blocks.

A Filestore is a Blockstore wrapper. When I moved providing resposibilities to Blockstore/Pinner/MFS... I forgot the Filestore. In Particular, the Filestore writes all blocks to the Blockstore except those of FilestoreNode type, which are serialized and written to the datastore directly via the Filemanager. The data serialized is not the block itself, but a pointer to a filesystem object and an offset.

So FilestoreNodes escaped providing even when the blockstore is configured to provide... Not anymore.