bitswap/httpnet: improve "Connect"/testCid check (#1057)
* bitswap/httpnet: improve "Connect"/testCid check

Now that we have certain safeguards to disconnect from small providers, we can be more open to connect to places that fail the "testCid request". Particularly places that return 500 with known error messages that prove that they understand ipld and that they are a gateway.

* bitswap/httpnet: one more 500 error that should be considered a 404 on retrieval

* bitswap/httpnet: Do not attempt test GET request when HEAD is 429.

* bitswap/httpnet: log cooldown-period errors

* bitswap/httpnet: ensure same errors from bsnet and httpnet

* bitswap/httpnet: better detection of 404s

* bitswap/httpnet: typo

* bitswap/httpnet: address feedback

* docs: add godoc for isKnownNotFoundError with robustness rationale

explains why we accept 5xx with known IPLD error messages and references
https://specs.ipfs.tech/architecture/principles/#robustness

* update changelog

---------

Co-authored-by: Marcin Rataj <lidel@lidel.org>